🚚 Move all http logs components in common
